moved burn-mark-sprite-supplier
