+++ engine/burn-mark.go
@@ -1,11 +1,14 @@
 package engine
 
-import "retro-carnage/engine/geometry"
+import (
+	"retro-carnage/engine/geometry"
+	"retro-carnage/engine/graphics"
+)
 
 // BurnMark is a mark on the ground caused by an explosion. They don't disappear.
 type BurnMark struct {
 	position       *geometry.Rectangle
-	SpriteSupplier *BurnMarkSpriteSupplier
+	SpriteSupplier *graphics.BurnMarkSpriteSupplier
 }
 
 func (bm *BurnMark) Position() *geometry.Rectangle {
