+++ engine/graphics/burn-mark-sprite-supplier.go
@@ -1,9 +1,8 @@
-package engine
+package graphics
 
 import (
 	"retro-carnage/assets"
 	"retro-carnage/engine/geometry"
-	"retro-carnage/engine/graphics"
 )
 
 type BurnMarkSpriteSupplier struct{}
@@ -12,10 +11,10 @@ const (
 	burnMarkSprite = "images/environment/burn-mark.png"
 )
 
-func (supplier *BurnMarkSpriteSupplier) Sprite() *graphics.SpriteWithOffset {
+func (supplier *BurnMarkSpriteSupplier) Sprite() *SpriteWithOffset {
 	var sprite = assets.SpriteRepository.Get(burnMarkSprite)
 	var offset = geometry.Point{X: 0, Y: 55}
-	return &graphics.SpriteWithOffset{
+	return &SpriteWithOffset{
 		Offset: offset,
 		Source: burnMarkSprite,
 		Sprite: sprite,
